Move to separate components as recommended by Dagger
