--- app-partial-users/src/main/java/com/jraska/github/client/users/SimpleRetrofitHolder.kt
@@ -1,10 +1,0 @@
-package com.jraska.github.client.users
-
-import com.jraska.github.client.HasRetrofit
-import retrofit2.Retrofit
-
-class SimpleRetrofitHolder(private val retrofit: Retrofit) : HasRetrofit {
-  override fun retrofit(): Retrofit {
-    return retrofit
-  }
-}
