+++ app-partial-users/src/main/java/com/jraska/github/client/users/UsersAppComponent.kt
@@ -1,21 +1,23 @@
 package com.jraska.github.client.users
 
 import android.app.Application
+import android.content.Context
 import androidx.lifecycle.ViewModelProvider
 import com.facebook.drawee.backends.pipeline.Fresco
 import com.jakewharton.threetenabp.AndroidThreeTen
-import com.jraska.github.client.CoreComponent
-import com.jraska.github.client.HasRetrofit
-import javax.inject.Singleton
+import com.jraska.github.client.FakeCoreModule
 import com.jraska.github.client.WebLinkLauncher
 import com.jraska.github.client.core.android.CoreAndroidModule
 import com.jraska.github.client.core.android.OnAppCreate
+import com.jraska.github.client.http.HttpModule
 import com.jraska.github.client.navigation.deeplink.DeepLinkNavigationModule
+import dagger.BindsInstance
 import dagger.Component
 import dagger.Module
 import dagger.Provides
 import dagger.multibindings.IntoSet
 import okhttp3.HttpUrl
+import javax.inject.Singleton
 
 @Singleton
 @Component(
@@ -23,16 +25,19 @@ import okhttp3.HttpUrl
     AppModule::class,
     CoreAndroidModule::class,
     DeepLinkNavigationModule::class,
-    UsersModule::class],
-  dependencies = [
-    HasRetrofit::class,
-    CoreComponent::class
-  ]
+    FakeCoreModule::class,
+    HttpModule::class,
+    UsersModule::class]
 )
 interface UsersAppComponent {
   fun onAppCreateActions(): Set<OnAppCreate>
 
   fun viewModelFactory(): ViewModelProvider.Factory
+
+  @Component.Factory
+  interface Factory {
+    fun create(@BindsInstance context: Context): UsersAppComponent
+  }
 }
 
 @Module
