+++ app/build.gradle
@@ -143,6 +143,7 @@ dependencies {
   androidTestImplementation 'com.airbnb.okreplay:espresso:1.6.0'
   androidTestImplementation 'com.squareup.rx.idler:rx2-idler:0.11.0'
   androidTestImplementation project(':core-testing')
+  kaptAndroidTest rootProject.ext.daggerAnnotationProcessor
 }
 
 apply plugin: 'com.jraska.module.graph.assertion'
