--- core-api/src/main/java/com/jraska/github/client/HasRetrofit.kt
@@ -1,7 +1,0 @@
-package com.jraska.github.client
-
-import retrofit2.Retrofit
-
-interface HasRetrofit {
-  fun retrofit(): Retrofit
-}
