+++ core-testing/build.gradle
@@ -1,6 +1,12 @@
 apply plugin: 'java-library'
 apply plugin: 'kotlin'
+apply plugin: 'kotlin-kapt'
 
 dependencies {
   implementation project(':core-api')
+
+  kapt rootProject.ext.daggerAnnotationProcessor
+  implementation rootProject.ext.dagger
+
+  implementation 'io.reactivex.rxjava2:rxjava:2.2.20'
 }
