+++ core-testing/src/main/java/com/jraska/github/client/FakeCoreModule.kt
@@ -3,25 +3,32 @@ package com.jraska.github.client
 import com.jraska.github.client.analytics.AnalyticsProperty
 import com.jraska.github.client.analytics.EventAnalytics
 import com.jraska.github.client.logging.CrashReporter
+import dagger.Module
+import dagger.Provides
 
-class FakeCoreComponent : CoreComponent {
+@Module
+object FakeCoreModule {
   val eventAnalytics = Fakes.recordingAnalytics()
   val config = Fakes.config()
   val analyticProperty = Fakes.recordingAnalyticsProperty()
 
-  override fun analytics(): EventAnalytics {
+  @Provides
+  fun analytics(): EventAnalytics {
     return eventAnalytics
   }
 
-  override fun analyticsProperty(): AnalyticsProperty {
+  @Provides
+  fun analyticsProperty(): AnalyticsProperty {
     return analyticProperty
   }
 
-  override fun crashReporter(): CrashReporter {
-    return Fakes.emptyCrashReporter()
+  @Provides
+  fun crashReporter(): CrashReporter {
+    return EmptyCrashReporter
   }
 
-  override fun config(): Config {
+  @Provides
+  fun config(): Config {
     return config
   }
 }
