+++ core-testing/src/main/java/com/jraska/github/client/Fakes.kt
@@ -26,10 +26,6 @@ object Fakes {
     return AppSchedulers(Schedulers.trampoline(), Schedulers.trampoline(), Schedulers.trampoline())
   }
 
-  fun emptyCrashReporter(): CrashReporter {
-    return EmptyCrashReporter
-  }
-
   fun recordingAnalyticsProperty(): RecordingAnalyticsProperty {
     return RecordingAnalyticsProperty()
   }
