+++ core/build.gradle
@@ -26,12 +26,19 @@ dependencies {
   implementation 'androidx.annotation:annotation:1.1.0'
   implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
 
+
   kapt rootProject.ext.daggerAnnotationProcessor
   implementation rootProject.ext.dagger
 
   implementation 'io.reactivex.rxjava2:rxandroid:2.1.1'
 
   implementation 'com.jakewharton.timber:timber:4.7.1'
+
+  implementation rootProject.ext.retrofit
+  implementation rootProject.ext.retrofitGsonConverter
+  implementation rootProject.ext.retrofitRxJavaAdapter
+  implementation 'com.google.code.gson:gson:2.8.6'
+  implementation rootProject.ext.okHttpLoggingInterceptor
 
   testImplementation 'junit:junit:4.13.1'
   testImplementation 'org.assertj:assertj-core:3.18.0'
